renaming routers to routes for consistency
